+++ CHANGELOG.md
@@ -1,1 +1,25 @@
 # Changelog
+
+<!-- 0.1.7 START -->
+
+## 0.1.7 (2022-02-)
+
+### Features and enhancements
+
+* Add [Handlebars](https://github.com/handlebars-lang/handlebars.js) template engine.
+* Add generated [@kubernetes/client-node](https://www.npmjs.com/package/@kubernetes/client-node) model to the source code
+* Add `kubectl get pod -o wide`-like column evaluating to Handlebars template engine.
+
+### Bug fixes
+
+<!-- 0.1.7 END -->
+
+<!-- 0.1.6 START -->
+
+## 0.1.6 (2022-02-01)
+
+### Features and enhancements
+
+### Bug fixes
+
+<!-- 0.1.6 END -->